+++ src/i18n/langs/en.json
@@ -1292,7 +1292,7 @@
     "FileTransfer": "File transfer",
     "OnlineSession": "Online session",
     "CreateCommandStorage": "Create command storage",
-    "CreateReplayStorage": "Create replay storage",
+    "CreateReplayStorage": "Create object storage",
     "Dashboard": "Dashboard",
     "DatabaseApp": "Database Apps",
     "DatabaseAppCreate": "Database app create",
@@ -1364,7 +1364,7 @@
     "ApplicationPermissionDetail": "Application permission detail",
     "ApplicationPermissionUpdate": "Application permission update",
     "RemoteAppUpdate": "Remote app update",
-    "ReplayStorageUpdate": "Replay storage update",
+    "ReplayStorageUpdate": "Replay object update",
     "Detail": "Detail",
     "Activity": "Activity",
     "SessionOffline": "Sessions offline",
@@ -1542,7 +1542,7 @@
     "remoteAddr": "Remote addr",
     "replay": "replay",
     "replaySession": "Replay session",
-    "replayStorage": "Replay storage",
+    "replayStorage": "Object storage",
     "riskLevel": "Risk level",
     "session": "Session",
     "sshPort": "SSH port",
