+++ src/i18n/langs/ja.json
@@ -1297,7 +1297,7 @@
     "OnlineSession": "オンラインセッション",
     "OnlineUserDevices": "オンラインデバイス",
     "CreateCommandStorage": "コマンドストアの作成",
-    "CreateReplayStorage": "録画ストレージの作成",
+    "CreateReplayStorage": "オブジェクトストレージを作成する",
     "Dashboard": "ダッシュボード",
     "DatabaseApp": "データベース",
     "DatabaseAppCreate": "データベースアプリケーションの作成",
@@ -1368,7 +1368,7 @@
     "ApplicationPermissionDetail": "アプリライセンスの詳細",
     "ApplicationPermissionUpdate": "アプリ認可ルールの更新",
     "RemoteAppUpdate": "リモートアプリケーションの更新",
-    "ReplayStorageUpdate": "録画保存の更新",
+    "ReplayStorageUpdate": "オブジェクトストレージを更新する",
     "Detail": "詳細",
     "SessionOffline": "履歴セッション",
     "SessionOnline": "オンラインセッション",
@@ -1535,7 +1535,7 @@
     "remoteAddr": "リモートアドレス",
     "replay": "リプレイ",
     "replaySession": "再生セッション",
-    "replayStorage": "録画保存",
+    "replayStorage": "オブジェクトストレージ",
     "riskLevel": "リスクレベル",
     "session": "会話",
     "sshPort": "SSHポート",
