+++ src/views/settings/Storage/ReplayStorage.vue
@@ -1,57 +1,48 @@
 <template>
-  <ListTable ref="ListTable" :header-actions="commandActions" :table-config="commandTableConfig" />
+  <ListTable ref="ListTable" :header-actions="replayActions" :table-config="replayTableConfig" />
 </template>
-
 <script>
-import ListTable from '@/components/Table/ListTable'
-import { SetToDefaultCommandStorage, TestCommandStorage } from '@/api/sessions'
+import ListTable from '@/components/Table/ListTable/index.vue'
+import { SetToDefaultReplayStorage, TestReplayStorage } from '@/api/sessions'
+import { getReplayStorageOptions } from '@/views/sessions/const'
 
 export default {
-  name: 'CommandStorage',
+  name: 'ReplayStorage',
   components: {
     ListTable
   },
-  props: {
-    object: {
-      type: Object,
-      default: () => ({})
-    }
-  },
   data() {
     const vm = this
+    const storageOptions = getReplayStorageOptions()
     return {
-      commandActions: {
-        canCreate: this.$hasPerm('terminal.add_commandstorage'),
+      replayActions: {
+        canCreate: this.$hasPerm('terminal.add_replaystorage'),
         hasExport: false,
         hasImport: false,
         hasRefresh: true,
         hasMoreActions: false,
         moreCreates: {
           callback: (item) => {
-            this.$router.push({ name: 'CreateCommandStorage', query: { type: item.name }})
+            console.log(item)
+            this.$router.push({ name: 'CreateReplayStorage', query: { type: item.name.toLowerCase() }})
           },
-          dropdown: [
-            {
-              name: 'es',
-              title: 'Elasticsearch'
-            }
-          ]
+          dropdown: storageOptions
         }
       },
-      commandTableConfig: {
-        title: 'command',
-        url: '/api/v1/terminal/command-storages/',
+      replayTableConfig: {
+        url: '/api/v1/terminal/replay-storages/',
         permissions: {
-          resource: 'commandstorage'
+          app: 'terminal',
+          resource: 'replaystorage'
         },
+        columnsExclude: ['meta'],
+        columns: [
+          'name', 'type', 'comment', 'is_default', 'actions'
+        ],
         columnsShow: {
-          min: ['name', 'type', 'actions'],
-          default: ['name', 'type', 'comment', 'is_default', 'actions']
+          min: ['name', 'type', 'actions']
         },
         columnsMeta: {
-          comment: {
-            sortable: 'custom'
-          },
           name: {
             formatter: function(row) {
               return row.name
@@ -65,26 +56,31 @@ export default {
             align: 'center',
             width: '100px'
           },
+          comment: {
+            sortable: 'custom'
+          },
           actions: {
             formatterArgs: {
-              canUpdate: function({ row }) {
-                return (row.name !== 'default' && row.name !== 'null' && vm.$hasPerm('terminal.change_commandstorage'))
-              },
               onUpdate: function({ row }) {
-                this.$router.push({ name: 'CommandStorageUpdate', params: { id: row.id }, query: { type: row.type.value }})
+                this.$router.push({ name: 'ReplayStorageUpdate', params: { id: row.id }, query: { type: row.type.value }})
+              },
+              canUpdate: function({ row }) {
+                return (
+                  row.name !== 'default' && row.name !== 'null' && vm.$hasPerm('terminal.change_replaystorage')
+                )
               },
               canDelete: function({ row }) {
-                return (row.name !== 'default' && row.name !== 'null' && vm.$hasPerm('terminal.delete_commandstorage'))
+                return (row.name !== 'default' && row.name !== 'null' && vm.$hasPerm('terminal.delete_replaystorage'))
               },
               hasClone: false,
               extraActions: [
                 {
                   name: 'test',
                   title: this.$t('sessions.test'),
+                  can: vm.$hasPerm('terminal.view_replaystorage'),
                   type: 'primary',
-                  can: vm.$hasPerm('terminal.view_commandstorage'),
                   callback: function({ row, col, cellValue, reload }) {
-                    TestCommandStorage(row.id).then(data => {
+                    TestReplayStorage(row.id).then(data => {
                       if (!data['is_valid']) {
                         this.$message.error(data.msg)
                       } else {
@@ -96,10 +92,10 @@ export default {
                 {
                   name: 'set_to_default',
                   title: this.$t('sessions.SetToDefault'),
+                  can: this.$hasPerm('terminal.change_replaystorage'),
                   type: 'primary',
-                  can: vm.$hasPerm('terminal.change_commandstorage'),
                   callback: function({ row, col, cellValue, reload }) {
-                    SetToDefaultCommandStorage(row.id).then(data => {
+                    SetToDefaultReplayStorage(row.id).then(data => {
                       vm.$refs.ListTable.reloadTable()
                       this.$message.success(this.$tc('sessions.SetSuccess'))
                     }).catch(() => {
@@ -115,12 +111,8 @@ export default {
     }
   },
   methods: {
-    createEs() {
-      this.$router.push({ name: 'CreateCommandStorage', query: { type: 'es' }})
-    }
   }
 }
-
 </script>
 
 <style scoped>
