+++ src/views/settings/Storage/ReplayStorageCreateUpdate.vue
@@ -9,7 +9,7 @@
 <script>
 import { GenericCreateUpdatePage } from '@/layout/components'
 import { STORAGE_TYPE_META_MAP } from '@/views/sessions/const'
-import { UpdateToken } from '@/components/Form/FormFields'
+import { UploadSecret } from '@/components/Form/FormFields'
 import { encryptPassword } from '@/utils/crypto'
 
 export default {
@@ -21,7 +21,7 @@ export default {
     const storageType = this.$route.query.type || 's3'
     const storageTypeMeta = STORAGE_TYPE_META_MAP[storageType] || {}
     return {
-      successUrl: { name: 'TerminalSetting', params: { activeMenu: 'RelayStorage' }},
+      successUrl: { name: 'Storage', params: { activeMenu: 'RelayStorage' }},
       url: `/api/v1/terminal/replay-storages/`,
       initial: {
         type: storageType,
@@ -49,8 +49,15 @@ export default {
         meta: {
           fields: storageTypeMeta.meta,
           fieldsMeta: {
-            SECRET_KEY: {
-              component: UpdateToken
+            SFTP_PASSWORD: {
+              hidden: (formValue) => formValue.STP_SECRET_TYPE !== 'password'
+            },
+            STP_PRIVATE_KEY: {
+              component: UploadSecret,
+              hidden: (formValue) => formValue.STP_SECRET_TYPE !== 'ssh_key'
+            },
+            STP_PASSPHRASE: {
+              hidden: (formValue) => formValue.STP_SECRET_TYPE !== 'ssh_key'
             }
           }
         },
@@ -59,7 +66,7 @@ export default {
         }
       },
       cleanFormValue(values) {
-        const encryptedFields = ['SECRET_KEY', 'ACCOUNT_KEY']
+        const encryptedFields = ['SFTP_PASSWORD', 'STP_PASSPHRASE']
         const meta = values.meta
         for (const item of encryptedFields) {
           const val = meta[item]
@@ -70,6 +77,8 @@ export default {
         return values
       }
     }
+  },
+  mounted() {
   }
 }
 </script>
